% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getAllPublications.R
\name{getAllPublications}
\alias{getAllPublications}
\title{Gets the list of publications on ptmdb}
\usage{
getAllPublications(db)
}
\arguments{
\item{db}{An object of the class \code{\link{MySQLConnection}}}
}
\value{
Produces a dataframe containing each of the publications in the database
}
\description{
The publications available in a given database are listed ordered by publication date
}
\examples{
dbConnection <- ptmdbConnect(configFile)
publications <- getAllPublications(db)
}

