% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptmdbConnect.R
\name{ptmdbConnect}
\alias{ptmdbConnect}
\title{Connects to ptmdb database}
\usage{
ptmdbConnect(...)
}
\arguments{
\item{...}{Arguments necessary to connect to the database. The arguments are inherited from the \code{\link{dbConnect}} function (see example)}
}
\value{
Produces an object of the class \code{\link{MySQLConnection}}
}
\description{
This function can be used to directly connect to the database using a configuration File.
}
\examples{
dbConnection <- ptmdbConnect(user=DBUSER, password=DBPASS, host=DBHOST, dbname=DATABASE, port=DBPORT)
}

