% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ptmdbR.R
\docType{package}
\name{ptmdbR}
\alias{ptmdbR}
\alias{ptmdbR-package}
\title{Useful functions for working with ptmdb}
\description{
ptmdbR provides a toolbox with some of the most useful functions necessary to interact with ptmdb database.
}
\details{
Database connection

Among many other features this package allows to establish a connection that can be used for querying the database using the \code{\link{ptmdbConnect}} function.

...
}
\references{
Coming soon
}

